requirements.txt
flask==1.1.2
sqlalchemy==1.4.12
Flask-SQLAlchemy==2.5.1
marshmallow==3.11.1
requests==2.24.0
flask-marshmallow==0.14.0
python-magic==0.4.18